Create `LinksTable` component
* Create `LinksTable` component (apps/client/src/components/ui/template/LinksTable/LinksTable.tsx, apps/client/src/components/ui/template/LinksTable/index.ts);
* Update `Home` page (apps/client/src/components/views/home/Home/Home.tsx).
