+++ apps/client/src/components/ui/template/LinksTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LinksTable";
